Creating a service to talk with auth0
