+++ src/modules/common/auth0.service.ts
@@ -1,0 +1,55 @@
+import { Injectable } from '@nestjs/common';
+import fetch from 'node-fetch';
+import Config from '../../config';
+import { response } from 'express';
+
+@Injectable()
+export class Auth0Service {
+
+  // Get an access token for the Auth0 Admin API
+  async getAdminAccessToken() {
+    const options = {
+      method: 'POST',
+      headers: { 'content-type': 'application/json' },
+      body: JSON.stringify({
+        client_id: Config.auth0.backend.CLIENT_ID,
+        client_secret: Config.auth0.backend.CLIENT_SECRET,
+        audience: `https://${Config.auth0.backend.DOMAIN}/api/v2/`,
+        grant_type: 'client_credentials',
+      }),
+    };
+
+    const response = await fetch(`https://${Config.auth0.backend.DOMAIN}/oauth/token`, options);
+    const json = await response.json();
+
+    return json;
+  }
+
+  // Get the user's profile from auth0
+  async getUserProfile(accessToken, userID) {
+    const options = {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    };
+
+    const response = await fetch(`https://${Config.auth0.backend.DOMAIN}/api/v2/users/${userID}`, options);
+    const json = await response.json();
+
+    return json;
+  }
+
+  // Deletes a user from auth0
+  async deleteUser(accessToken: string, userID: string) {
+    const options = {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    };
+
+    const response = await fetch(`https://${Config.auth0.backend.DOMAIN}/api/v2/users/${userID}`, options);
+
+    return response;
+  }
+}
