+++ src/modules/common/common.module.ts
@@ -1,4 +1,5 @@
 import { Module } from '@nestjs/common';
+import { Auth0Service } from './auth0.service';
 import { UsersService } from './users.service';
 import { commonProviders } from './common.providers';
 import { DatabaseModule } from '../../database/database.module';
@@ -6,7 +7,7 @@ import { MentorsService } from './mentors.service';
 
 @Module({
   imports: [DatabaseModule],
-  providers: [MentorsService, UsersService, ...commonProviders],
-  exports: [MentorsService, UsersService, ...commonProviders],
+  providers: [MentorsService, UsersService, Auth0Service, ...commonProviders],
+  exports: [MentorsService, UsersService, Auth0Service, ...commonProviders],
 })
 export class CommonModule { }
